move debug-handler pkg into jimmhttp
