+++ internal/jimmhttp/debug_handler_test.go
@@ -1,5 +1,5 @@
 // Copyright 2024 Canonical.
-package debugapi_test
+package jimmhttp_test
 
 import (
 	"context"
@@ -14,15 +14,15 @@ import (
 	qt "github.com/frankban/quicktest"
 	"github.com/go-chi/chi/v5"
 
-	"github.com/canonical/jimm/v3/internal/debugapi"
 	"github.com/canonical/jimm/v3/internal/errors"
+	"github.com/canonical/jimm/v3/internal/jimmhttp"
 	"github.com/canonical/jimm/v3/version"
 )
 
-func setupHandlerAndRecorder(c *qt.C, startTime debugapi.StatusCheck, path string) *httptest.ResponseRecorder {
-	r := (&debugapi.DebugHandler{
+func setupHandlerAndRecorder(c *qt.C, startTime jimmhttp.StatusCheck, path string) *httptest.ResponseRecorder {
+	r := (&jimmhttp.DebugHandler{
 		Router: chi.NewRouter(),
-		StatusChecks: map[string]debugapi.StatusCheck{
+		StatusChecks: map[string]jimmhttp.StatusCheck{
 			"start_time": startTime,
 		},
 	}).Routes()
@@ -37,7 +37,7 @@ func setupHandlerAndRecorder(c *qt.C, startTime debugapi.StatusCheck, path strin
 func TestDebugInfo(t *testing.T) {
 	c := qt.New(t)
 
-	rr := setupHandlerAndRecorder(c, debugapi.ServerStartTime, "/info")
+	rr := setupHandlerAndRecorder(c, jimmhttp.ServerStartTime, "/info")
 
 	resp := rr.Result()
 	defer resp.Body.Close()
@@ -50,10 +50,10 @@ func TestDebugInfo(t *testing.T) {
 func TestDebugStatus(t *testing.T) {
 	c := qt.New(t)
 	ctx := context.Background()
-	startTime, err := debugapi.ServerStartTime.Check(ctx)
+	startTime, err := jimmhttp.ServerStartTime.Check(ctx)
 	c.Assert(err, qt.IsNil)
 
-	rr := setupHandlerAndRecorder(c, debugapi.ServerStartTime, "/status")
+	rr := setupHandlerAndRecorder(c, jimmhttp.ServerStartTime, "/status")
 
 	resp := rr.Result()
 	defer resp.Body.Close()
@@ -66,7 +66,7 @@ func TestDebugStatus(t *testing.T) {
 	err = json.Unmarshal(buf, &v)
 	c.Assert(err, qt.IsNil)
 
-	c.Check(v["start_time"]["Name"], qt.Equals, debugapi.ServerStartTime.Name())
+	c.Check(v["start_time"]["Name"], qt.Equals, jimmhttp.ServerStartTime.Name())
 	c.Check(v["start_time"]["Value"], qt.Equals, startTime.(time.Time).Format(time.RFC3339Nano))
 	c.Check(v["start_time"]["Passed"], qt.Equals, true)
 }
@@ -74,7 +74,7 @@ func TestDebugStatus(t *testing.T) {
 func TestDebugStatusStatusError(t *testing.T) {
 	c := qt.New(t)
 
-	rr := setupHandlerAndRecorder(c, debugapi.MakeStatusCheck("Test", func(context.Context) (interface{}, error) {
+	rr := setupHandlerAndRecorder(c, jimmhttp.MakeStatusCheck("Test", func(context.Context) (interface{}, error) {
 		return nil, errors.E("test error")
 	}), "/status")
